Extract todo to seperate module
