+++ index.html
@@ -14,8 +14,8 @@
     <div class="container mx-auto font-sans text-gray-900 p-4 max-w-screen-lg text-sm md:text-base">
 
         <header class="grid grid-cols-2 bg-green-300 rounded md:h-20 md:shadow-xl">
-            <h1 class="text-5xl font-bold pl-2">Pompom</h1>
-            <div class="grid justify-end items-center pr-2">
+            <h1 class="text-5xl font-bold pl-4">Pompom</h1>
+            <div class="grid justify-end items-center pr-4">
                 <button id="btn-login" class="hidden bg-transparent h-8 hover:bg-green-500 text-green-700 font-semibold uppercase hover:text-white px-2 border border-green-500 hover:border-transparent rounded text-xs md:text-sm shadow-md">Log in</button>
                 <button id="btn-logout" class="hidden bg-transparent h-8 hover:bg-green-500 text-green-700 font-semibold uppercase hover:text-white px-2 border border-green-500 hover:border-transparent rounded text-xs md:text-sm shadow-md">Log out</button>
             </div>
@@ -32,13 +32,13 @@
                     <button data-reset class="bg-transparent hover:bg-green-500 text-green-700 font-semibold uppercase hover:text-white py-2 px-4 border border-green-500 hover:border-transparent rounded text-xs md:text-sm shadow-md">Reset</button>
                 </div>
             </section>
-            <section class="grid col-span-2 md:col-span-1 content-start p-4 shadow-md relative bg-green-300 rounded md:pt-8 md:-mt-8">
+            <section class="grid col-span-2 md:col-span-1 content-start p-4 shadow-md relative bg-green-300 rounded md:pt-8 md:-mt-4 md:shadow-xl">
                 <h2 class="text-xl font-bold text-center uppercase">Todo List <i id="task-add-button" class="material-icons text-green-700" title="Add">note_add</i></h2>
                 <div class="italic text-sm text-center">Select a task below to use as current item</div>
                 <div data-task-list class="pt-4 px-4"></div>
                 
             </section>
-            <section class="grid col-span-2 md:col-span-1 content-start p-4 shadow-md bg-green-300 rounded md:pt-8 md:-mt-8">
+            <section class="grid col-span-2 md:col-span-1 content-start p-4 shadow-md bg-green-300 rounded md:pt-8 md:-mt-4 md:shadow-xl">
                 <h2 class="text-xl font-bold text-center uppercase">Completed Tasks</h2>
                 <div class="italic text-sm text-center">Tasks completed will be listed below</div>
                 <div data-complete-tasks class="pt-4 px-4"></div>
